+++ Backend/routes/create.routes.js
@@ -2,7 +2,7 @@ var SpotifyWebApi = require('spotify-web-api-node');
 var MongoClient = require('mongodb').MongoClient;
 var assert = require('assert');
 var uri = "mongodb+srv://jumarogu:rooster@cluster0-vxv1v.mongodb.net/PWF?retryWrites=true";
-exports.sayHi = (req, res) => {
+exports.createPlaylist = (req, res) => {
     
     if(req.body.access_token != null) {
         let access_token = req.body.access_token;
@@ -10,7 +10,7 @@ exports.sayHi = (req, res) => {
         var spotifyApi = new SpotifyWebApi({
             clientId: 'be2a413e2bbd402db45432d7ccdf0199',
             clientSecret: 'aa634b8aaac3420ea1056e6a533d939e',
-            redirectUri: 'http://localhost:4200/create'
+            redirectUri: 'http://localhost:4200/show-code'
           });
         spotifyApi.setAccessToken(access_token);
